# mason/emitter: emit pkgs in parallel