Depends on https://github.com/serpent-os/moss-deps/pull/4

Quick n' Dirty Benchmarks
glibc : 28s -> 8s
curl  : 1.17s -> 668ms
nano  : 451ms -> 311ms

Sub task of #41.